Project Commit
This commit has most of the planned features for this project. As of this commit
the planned trading function has not been implemented.  That feature turned out
to be much more complicated than anticipated.  I am turning in the current state
of my project that meets the project requirements.  I have done a lot work on the
trading feature but it is not ready yet.  Some of the work can be seen in the code.
There is a trade controller with some views as well as a trade table and a trade model.